requirements.txt
requests
Pillow
ansicolors
numpy
pysocks
stem
tflite-runtime[auto-captcha]
uvicorn
fastapi
starlette
